fix(review): resolve unresolved #1789 review comments
- TRIE.Train: restore eval mode in a finally so a throwing TrainWithTape can't
  strand the model in training mode (matches DocGCN/PICK/LayoutLMv2).
- LayoutLMv2.DistributeLayers: throw if the layer count drifts below the
  visual+text role split instead of silently misclassifying layers.
- OpticalFlowBase: compare Layers.Count against (long)numLayers + 2 so a huge
  serialized numLayers can't wrap the int addition negative and bypass the guard.
- VideoInpaintingBase.DenormalizeInpaintFrames: skip the hard pixel-range clamp
  during training (PostprocessOutput runs inside ForwardForTraining) so the
  reconstruction gradient for out-of-range pixels isn't zeroed; still clamp for
  inference output.
- Conv3DLayer / MultiHeadAttentionLayer: use the shared WeightsAlreadyAllocated
  helper instead of duplicated inline rank/shape checks.
- DepthwiseSeparableConvolutionalLayer: validate all three parameter tensors in
  the idempotent guard, not just the depthwise kernels, so a partial restore
  can't skip allocating the others.
- LayerHelper: make the DeepFilterNet layer-count constants internal (same-assembly
  plumbing, matches ODISEDecoderStageLayerCount).

+++ src/Video/OpticalFlowBase.cs
@@ -254,7 +254,10 @@ public abstract class OpticalFlowBase<T> : VideoNeuralNetworkBase<T>
         // at Layers[numLayers + 1]. The ctor already enforces numLayers > 0.
         if (numLayers < 0)
             throw new InvalidDataException($"{modelName} serialized layer count {numLayers} is invalid.");
-        if (Layers.Count < numLayers + 2)
+        // Compare in long space: numLayers + 2 as unchecked int wraps negative for numLayers near
+        // int.MaxValue, bypassing this guard and throwing a raw IndexOutOfRangeException at
+        // Layers[numLayers + 1] below. (long)numLayers + 2 cannot overflow for any int input.
+        if (Layers.Count < (long)numLayers + 2)
             throw new InvalidDataException(
                 $"{modelName} serialized layer count {Layers.Count} is too small for {numLayers} processing blocks.");
         featureExtract = Layers[0] as ConvolutionalLayer<T>
